+++ backend/src/main/java/com/webappsbusters/parcelmanagement/mapper/OrikaMapper.java
@@ -1,4 +1,4 @@
-package com.webappsbusters.parcelmanagement;
+package com.webappsbusters.parcelmanagement.mapper;
 
 import com.webappsbusters.parcelmanagement.domain.Parcel;
 import com.webappsbusters.parcelmanagement.domain.ParcelDto;
@@ -7,14 +7,14 @@ import ma.glasnost.orika.impl.ConfigurableMapper;
 import org.springframework.stereotype.Component;
 
 @Component
-public class OrikaAttempt extends ConfigurableMapper {
+public class OrikaMapper extends ConfigurableMapper {
     @Override
     protected void configure(MapperFactory factory) {
         factory.classMap(Parcel.class, ParcelDto.class)
                 .byDefault()
                 .register();
-        factory.classMap(size.class, sizeDto.class)
-                .byDefault()
-                .register();
+        //factory.classMap(size.class, sizeDto.class)
+        //       .byDefault()
+        //        .register();
     }
 }
